refactor(server): split runs handler and runs test file into modular sub-modules (warren-6566)
