+++ .mulch/expertise/server.jsonl
@@ -81,3 +81,4 @@
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-23T15:14:30.401Z","evidence":{"commit":"092800d8ebe001edeb3188803cc9bcb26e47a6a7"},"content":"GET /analytics/cost (src/server/handlers.ts listCostAnalyticsHandler, warren-cf63 / pl-b0c0 step 6) returns a single payload with totals + all eight breakdowns (date/project/plan/plot/run/agent/model/provider). Query params: from/to (ISO8601 datetime, validated via new Date()), projectId. Defaults to last 30 days when neither bound is set. Handler pipeline: RunsRepo.listForAnalytics({projectId,from,to}) -> hydrateRunsUsage -> PlanRunsRepo.resolvePlanForRunIds -> extractProviderModel (reads renderedAgentJson.frontmatter via readProviderFrontmatter) -> buildCostAnalytics. Added '/analytics' to API_PREFIXES AND src/preview/proxy.ts WARREN_API_PATH_PREFIXES (warren-63e1 parity test enforces this).","id":"mx-dbd77e"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-27T22:48:05.083Z","evidence":{"commit":"3cfbdbb751409900a4ca73ba873071a9a82b13ed"},"relates_to":["mx-5009ad"],"content":"src/server/main.ts (warren-8d3d / pl-9088 step 10) is split into src/server/main/ subdirectory: utils.ts (defaultSpawn + closeDatabase + redactDbUrl + resolvePgPoolMax + parseTrueEnv + parseIntEnv), logging.ts (6 pino→narrow logger adapters: bridge/probe/scheduler/planRun/pause/previewEviction), preview-wiring.ts (createPreviewAuthAndProxy — R-19/SPEC §11.L token × mode × host matrix), deps.ts (buildServerDeps — ServerDeps literal + plotAggregator), index.ts (bootServer orchestrator + CLI entry + public type re-exports, 472 lines). Test files split 1:1 alongside (utils.test.ts/logging.test.ts/preview-wiring.test.ts + the original resolvePgPoolMax suite at main/index.test.ts) so the extracted helpers regain direct coverage — otherwise the split would have dropped lines coverage below the 90% floor since boot-only code lives in the 0%-executed bootServer body. Importers (src/cli/commands/serve.{ts,test.ts}, src/server/index.ts) re-point to main/index.ts. scripts/file-size-budgets.json entry for main.ts removed; biome.json noExcessiveCognitiveComplexity override updated main.ts -> main/index.ts (bootServer body still trips the threshold). The CLI entry shebang (import.meta.main block) lives in main/index.ts; bun run src/server/main/index.ts replaces the prior src/server/main.ts invocation in the Phase-12 supervisor exec path.","id":"mx-131efe"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-28T03:05:28.261Z","evidence":{"commit":"245190a440fa18c003d43464dcb4f4a1cce127e7"},"dir_anchors":["src/server/handlers"],"content":"plots/ index.ts is a composing route table for plots subdirectory, re-exporting the split handlers list, sync, workbench and remaining monolithic plots.ts handlers.","id":"mx-b7697e"}
+{"type":"convention","classification":"tactical","recorded_at":"2026-05-28T03:29:35.665Z","evidence":{"commit":"bf1bbc8819f28805972385f605bcb3512c92315b"},"dir_anchors":["src/server/handlers"],"content":"Split src/server/handlers/runs.ts into runs/ subdirectory domain modules (dispatch, lifecycle, events, preview, pause-resume) and runs.test.ts into modular test files (runs.dispatch.test, runs.interactive.test, runs.events.test) with shared test helpers (warren-6566)","id":"mx-cad464"}
