+++ src/server/handlers/runs/index.ts
@@ -1,0 +1,27 @@
+/**
+ * Composite export for the runs handlers domain (warren-6566 / pl-3255 step 3).
+ */
+
+export {
+	createRunHandler,
+	postRunMessageHandler,
+} from "./dispatch.ts";
+export {
+	asNdjsonStream,
+	bridgeAbort,
+	eventToNdjson,
+	streamRunEventsHandler,
+} from "./events.ts";
+export {
+	getRunHandler,
+	listCostAnalyticsHandler,
+	listRunsHandler,
+} from "./lifecycle.ts";
+export {
+	cancelRunHandler,
+	steerRunHandler,
+} from "./pause-resume.ts";
+export {
+	previewLoginHandler,
+	previewTeardownHandler,
+} from "./preview.ts";
