+++ src/server/handlers/runs/pause-resume.ts
@@ -1,0 +1,55 @@
+import type { MessagePriority } from "@os-eco/burrow-cli";
+import { cancelRun, steerRun } from "../../../runs/index.ts";
+import { jsonResponse } from "../../response.ts";
+import type { RouteHandler, ServerDeps } from "../../types.ts";
+import {
+	optionalString,
+	readJsonBody,
+	readJsonBodyOrEmpty,
+	requireParam,
+	requireString,
+} from "../index.ts";
+
+export function steerRunHandler(deps: ServerDeps): RouteHandler {
+	return async (ctx) => {
+		const id = requireParam(ctx, "id");
+		const body = await readJsonBody(ctx);
+		const result = await steerRun({
+			runId: id,
+			body: requireString(body, "body"),
+			repos: deps.repos,
+			burrowClientPool: deps.burrowClientPool,
+			broker: deps.broker,
+			...(optionalString(body, "priority") !== undefined
+				? { priority: optionalString(body, "priority") as MessagePriority }
+				: {}),
+			...(optionalString(body, "fromActor") !== undefined
+				? { fromActor: optionalString(body, "fromActor") as string }
+				: {}),
+			...(deps.now !== undefined ? { now: deps.now } : {}),
+		});
+		return jsonResponse(200, { message: result.message });
+	};
+}
+
+export function cancelRunHandler(deps: ServerDeps): RouteHandler {
+	return async (ctx) => {
+		const id = requireParam(ctx, "id");
+		const body = await readJsonBodyOrEmpty(ctx);
+		const reason = body !== null ? optionalString(body, "reason") : undefined;
+		const result = await cancelRun({
+			runId: id,
+			repos: deps.repos,
+			burrowClientPool: deps.burrowClientPool,
+			broker: deps.broker,
+			...(reason !== undefined ? { reason } : {}),
+			...(deps.now !== undefined ? { now: deps.now } : {}),
+			...(deps.autoOpenPr !== undefined ? { autoOpenPr: deps.autoOpenPr } : {}),
+		});
+		return jsonResponse(200, {
+			state: result.state,
+			alreadyTerminal: result.alreadyTerminal,
+			burrowRun: result.burrowRun,
+		});
+	};
+}
